Change User ID to an Integer.
Change UserRepository entity ID declaration to an Integer.

+++ user-domain/src/main/java/example/app/model/User.java
@@ -45,12 +45,12 @@ public class User {
 
 	@Id
 	@Setter(AccessLevel.PROTECTED)
-	private Long id;
+	private Integer id;
 
 	@lombok.NonNull
 	private String name;
 
-	public @NonNull User identifiedBy(@NonNull Long id) {
+	public @NonNull User identifiedBy(@NonNull Integer id) {
 		Assert.notNull(id, "ID is required");
 		setId(id);
 		return this;
